Move remaining unary ops
